+++ src/matrix.c
@@ -12,9 +12,9 @@
 
 // Matrix creation functions
 
-matrix_t * matrix_create(unsigned int rows, unsigned int columns)
+matrix_t * matrix_create(size_t rows, size_t columns)
 {
-    unsigned int i = 0;
+    size_t i = 0;
     matrix_t *matrix = malloc(sizeof(matrix_t));
     if (!matrix)
         goto failed_matrix;
@@ -32,7 +32,7 @@ matrix_t * matrix_create(unsigned int rows, unsigned int columns)
     }
     return matrix;
 failed_coeff_elt:
-    for (unsigned int j = 0; j < i; j++)
+    for (size_t j = 0; j < i; j++)
         free(matrix->coeff[j]);
     free(matrix->coeff);
 failed_coeff:
@@ -42,10 +42,10 @@ failed_matrix:
     return NULL;
 }
 
-matrix_t * matrix_identity(unsigned int n)
+matrix_t * matrix_identity(size_t n)
 {
     matrix_t * matrix = matrix_create(n, n);
-    for (unsigned int i = 0; i < n; i++)
+    for (size_t i = 0; i < n; i++)
         matrix->coeff[i][i] = 1;
     return matrix;
 }
@@ -53,9 +53,9 @@ matrix_t * matrix_identity(unsigned int n)
 matrix_t * matrix_copy(const matrix_t *matrix)
 {
     matrix_t *copy = matrix_create(matrix->rows, matrix->columns);
-    for (unsigned int i = 0; i < matrix->rows; i++) {
+    for (size_t i = 0; i < matrix->rows; i++) {
         #pragma omp simd
-        for (unsigned int j = 0; j < matrix->columns; j++) {
+        for (size_t j = 0; j < matrix->columns; j++) {
             copy->coeff[i][j] = matrix->coeff[i][j];
         }
     }
@@ -65,7 +65,7 @@ matrix_t * matrix_copy(const matrix_t *matrix)
 void matrix_free(matrix_t *matrix)
 {
     if(!sanity_check(matrix, __func__))return; 
-    for (unsigned int i = 0; i < matrix->rows; i++) {
+    for (size_t i = 0; i < matrix->rows; i++) {
         free(matrix->coeff[i]);
     }
     free(matrix->coeff);
@@ -81,8 +81,8 @@ matrix_t * matrix_transp_f(const matrix_t *matrix)
     if(!sanity_check((void *)matrix, __func__))return NULL; 
     matrix_t *transpose_matrix = matrix_create(matrix->columns, matrix->rows);
     #pragma omp parallel for
-    for (unsigned int i = 0; i < transpose_matrix->rows; i++)
-        for (unsigned int j = 0; j < transpose_matrix->columns; j++)
+    for (size_t i = 0; i < transpose_matrix->rows; i++)
+        for (size_t j = 0; j < transpose_matrix->columns; j++)
             transpose_matrix->coeff[i][j]=matrix->coeff[j][i];
     return transpose_matrix;
 }
@@ -96,9 +96,9 @@ matrix_t * matrix_add_f(const matrix_t *matrix1, const matrix_t *matrix2)
         return NULL;
     }
     matrix_t *add_matrix = matrix_create(matrix1->rows, matrix1->columns);
-    for (unsigned int i = 0; i < add_matrix->rows; i++) {
+    for (size_t i = 0; i < add_matrix->rows; i++) {
         #pragma omp simd
-        for (unsigned int j = 0; j < add_matrix->columns; j++) {
+        for (size_t j = 0; j < add_matrix->columns; j++) {
             add_matrix->coeff[i][j] = matrix1->coeff[i][j] + matrix2->coeff[i][j];
         }
     }
@@ -109,9 +109,9 @@ matrix_t * matrix_mult_scalar_f(const matrix_t *matrix, TYPE lambda)
 {
     if(!sanity_check((void *)matrix, __func__))return NULL;
     matrix_t *mult_matrix = matrix_create(matrix->rows, matrix->columns);
-    for (unsigned int i = 0; i < matrix->rows; i++) {
+    for (size_t i = 0; i < matrix->rows; i++) {
         #pragma omp simd
-        for (unsigned int j = 0; j < matrix->columns; j++) {
+        for (size_t j = 0; j < matrix->columns; j++) {
             mult_matrix->coeff[i][j]=lambda * matrix->coeff[i][j];
         }
     }
@@ -127,20 +127,20 @@ matrix_t * matrix_mult_f(const matrix_t *matrix1, const matrix_t *matrix2)
         return NULL;
     }
     int step = 16;
-   	unsigned int n = matrix1->rows, m = matrix2->columns;
+   	size_t n = matrix1->rows, m = matrix2->columns;
     matrix_t *mult = matrix_create(matrix1->columns, matrix2->columns);
     if(!sanity_check((void *)mult, __func__))return NULL; 
     matrix_t *columns = matrix_transp_f(matrix2);
     if(!sanity_check((void *)columns, __func__))return NULL; 
     #pragma omp parallel for
-    for (unsigned int i = 0; i < n; i+=step) {
+    for (size_t i = 0; i < n; i+=step) {
         int ie = n < i+step ? n : i+step;
-        for (unsigned int j = 0; j < m; j+=step) {
+        for (size_t j = 0; j < m; j+=step) {
             int je = m < j+step ? m : j+step;
             for (int ii = i; ii < ie; ii++){
                 for (int jj = j; jj < je; jj++){
                     TYPE sum = 0;
-                    for (unsigned int k = 0; k < n; k++)
+                    for (size_t k = 0; k < n; k++)
                         sum+=matrix1->coeff[ii][k]*columns->coeff[jj][k];
                     mult->coeff[ii][jj] += sum;
                 }
